Fix server arch test
